pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

# uv-specific configuration
[tool.uv]
# Use managed Python versions for better reproducibility
python-preference = "managed"

[project]
name = "templateheaven"
version = "0.1.0"
description = "Interactive template management for Template Heaven"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Template Heaven Team", email = "team@templateheaven.dev"},
]
maintainers = [
    {name = "Template Heaven Team", email = "team@templateheaven.dev"},
]
keywords = ["templates", "boilerplate", "project-starter", "cli", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "questionary>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "aiohttp>=3.9",
    # API dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psutil>=5.9.0",
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "pytest-asyncio>=1.2.0",
    "black>=23.0",
    "mypy>=1.0",
    "flake8>=6.0",
    "isort>=5.12",
]
ui = [
    "streamlit>=1.28",
]
github = [
    "gitpython>=3.1",
    "requests>=2.28",
]
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/template-heaven/templateheaven"
Documentation = "https://templateheaven.dev/docs"
Repository = "https://github.com/template-heaven/templateheaven.git"
Issues = "https://github.com/template-heaven/templateheaven/issues"
Changelog = "https://github.com/template-heaven/templateheaven/blob/main/CHANGELOG.md"

[project.scripts]
templateheaven = "templateheaven.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["templateheaven*"]

[tool.setuptools.package-data]
templateheaven = ["data/*.yaml", "data/templates/*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["templateheaven"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=templateheaven",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running",
    "external: marks tests that require external services",
]

[tool.coverage.run]
source = ["templateheaven"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
